+++ common/src/main/java/net/conczin/mca/client/render/VillagerRenderState.java
@@ -5,4 +5,5 @@ import net.minecraft.client.renderer.entity.state.UndeadRenderState;
 public class VillagerRenderState extends UndeadRenderState {
     public boolean isConverting;
     public float panicAnimationProgress;
+    public boolean mcaInventoryPreview;
 }